requirements.txt
anyio
cachetools
certifi
charset-normalizer
click
colorama
Flask
google-api-core
google-auth
google-cloud-vision
googleapis-common-protos
grpcio
grpcio-status
h11
httpcore
httpx
idna
itsdangerous
Jinja2
MarkupSafe
numpy
packaging
pandas
Pillow
proto-plus
protobuf
pyasn1
pyasn1-modules
python-dateutil
python-dotenv
python-telegram-bot
pytz
requests
rfc3986
rsa
six
sniffio
urllib3
Werkzeug
